# backend/docs/swagger.yaml
basePath: /
definitions:
...
  handlers.ApiErrorResponse:
    properties:
      error:
        type: string
    required:
    - error
    type: object
  handlers.ArtistIDRequest:
    properties:
      artist_id:
        example: 1
        type: integer
    required:
    - artist_id
    type: object
...
  handlers.CreateAlbumRequest:
    properties:
      artist_id:
        example: 1
        type: integer
      spotify_id:
        example: 4oDw9mW4Sro2zN1RHzlvOr
        type: string
      title:
        example: Abbey Road
        type: string
    required:
    - artist_id
    - title
    type: object
...
  handlers.MessageResponse:
    properties:
      message:
        type: string
    required:
    - message
    type: object
...
  handlers.UpdateAlbumRequest:
    properties:
      artist_id:
        example: 1
        type: integer
      comment:
        example: Classic album
        type: string
      cover_url:
        example: https://example.com/cover.jpg
        type: string
      listened:
        example: true
        type: boolean
      listened_at:
        example: "2024-01-15"
        type: string
      rating:
        example: 8
        type: integer
      spotify_id:
        example: 4oDw9mW4Sro2zN1RHzlvOr
        type: string
      title:
        example: Abbey Road
        type: string
      year:
        example: 1969
        type: integer
    required:
    - artist_id
    type: object
...
  models.Album:
    properties:
      artist_id:
        type: integer
      comment:
        type: string
      cover_url:
        type: string
      created_at:
        type: string
      id:
        type: integer
      listened:
        type: boolean
      listened_at:
        type: string
      rating:
        type: integer
      spotify_id:
        type: string
      title:
        type: string
      year:
        type: integer
    required:
    - artist_id
    - created_at
    - id
    - listened
    - title
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: A music management API with artists and albums
  title: Better Music API
  version: "1.0"
paths:
...
  /api/albums:
    get:
      description: Get all albums for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Album'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ApiErrorResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/handlers.ApiErrorResponse'
      security:
      - Bearer: []
      summary: Get all albums
      tags:
      - albums
    post:
      consumes:
      - application/json
      description: Create a new album for an artist
      parameters:
      - description: Album data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateAlbumRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Album'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/handlers.ApiErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ApiErrorResponse'
        "409":
          description: Album already exists
          schema:
            $ref: '#/definitions/handlers.ApiErrorResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/handlers.ApiErrorResponse'
      security:
      - Bearer: []
      summary: Create a new album
      tags:
      - albums
  /api/albums/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an album by ID
      parameters:
      - description: Album ID
        in: path
        name: id
        required: true
        type: integer
      - description: Artist ID
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.ArtistIDRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.MessageResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/handlers.ApiErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ApiErrorResponse'
        "404":
          description: Album or artist not found
          schema:
            $ref: '#/definitions/handlers.ApiErrorResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/handlers.ApiErrorResponse'
      security:
      - Bearer: []
      summary: Delete an album
      tags:
      - albums
    get:
      consumes:
      - application/json
      description: Get a specific album by ID
      parameters:
      - description: Album ID
        in: path
        name: id
        required: true
        type: integer
      - description: Artist ID
        in: query
        name: artist_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Album'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/handlers.ApiErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ApiErrorResponse'
        "404":
          description: Album or artist not found
          schema:
            $ref: '#/definitions/handlers.ApiErrorResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/handlers.ApiErrorResponse'
      security:
      - Bearer: []
      summary: Get a specific album
      tags:
      - albums
    put:
      consumes:
      - application/json
      description: Update an album's information
      parameters:
      - description: Album ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateAlbumRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.MessageResponse'
        "400":
          description: Invalid request or no fields provided
          schema:
            $ref: '#/definitions/handlers.ApiErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ApiErrorResponse'
        "404":
          description: Album or artist not found
          schema:
            $ref: '#/definitions/handlers.ApiErrorResponse'
        "409":
          description: Album already exists
          schema:
            $ref: '#/definitions/handlers.ApiErrorResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/handlers.ApiErrorResponse'
      security:
      - Bearer: []
      summary: Update an album
      tags:
      - albums
...
schemes:
- http
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"